Enhance scripts

+++ scripts/etsi_its_patches/denm_generate_makefile.bash
@@ -1,5 +1,6 @@
 #!/bin/bash
-set -x
+#set -e
+set -vx
 
 function f_exit {
     cd ${CURPWD}
@@ -26,7 +27,7 @@ then
     f_usage
 fi
 
-ATS_NAME=DENM
+ATS_NAME=Denm
 
 #CURPWD=`pwd`
 if [ ! "${PWD##*/}" == "objs" ]
@@ -84,29 +85,31 @@ then
     fi
 fi
 
-REFERENCES="LibCommon LibIts/Common LibIts/BTP LibIts/GeoNetworking LibIts/GeoNetworking LibIts/Ipv6OverGeoNetworking LibIts/Security LibIts/DENM"
+REFERENCES="LibCommon LibIts/Common LibIts/BTP LibIts/GeoNetworking LibIts/Ipv6OverGeoNetworking LibIts/Security LibIts/CAM LibIts/DENM"
 for i in ${REFERENCES}
 do
     # TTCN code
     for j in `find ${PATH_DEV_ITS}/src/$i/ttcn -type f -name "*.ttcn"`;
     do
-	ln -sf $j ../ttcn/`basename $j`
+	      ln -sf $j ../ttcn/`basename $j`
     done
-    # Include code
-#    if [ "`ls ${PATH_DEV_ITS}/src/$i/include`" != " " ]
-#    then
-#	for j in `find ${PATH_DEV_ITS}/src/$i/include -type f`;
-#	do
-#	    ln -sf $j ../include/`basename $j`
-#	done
-#    fi
-    # CC source code
-    if [ "`ls ${PATH_DEV_ITS}/src/$i/src`" != " " ]
+    # Include source code
+    files=`find ${PATH_DEV_ITS}/src/$i/include -type f`
+    if [ "${files}" != " " ]
     then
-	for j in `find ${PATH_DEV_ITS}/src/$i/src -type f"`;
-	do
-	    ln -sf $j ../src/`basename $j`
-	done
+	      for j in ${files};
+	      do
+	          ln -sf $j ../include/`basename $j`
+	      done
+    fi
+    # CC source code
+    files=`find ${PATH_DEV_ITS}/src/$i/src -type f`
+    if [ "${files}" != " " ]
+    then
+	      for j in ${files};
+	      do
+	          ln -sf $j ../src/`basename $j`
+	      done
     fi
 done
 
@@ -117,13 +120,13 @@ TTCN_FILES=`find .. -name '*.ttcn*'`
 if [ "${OSTYPE}" == "cygwin" ]
 then
     rm ../bin/*.exe ../lib/*.dll
-    compiler.exe -b -d -f -g -j -l -L -t -R -U none -x -X ${TTCN_FILES} ${ASN1_FILES} 2>&1 3>&1 | tee build.log
+    compiler.exe -d -e -f -g -j -l -L -t -R -U none -x -X ${TTCN_FILES} ${ASN1_FILES} 2>&1 3>&1 | tee build.log
     if [ "$?" == "1" ]
     then
         f_exit "Failed to compile ATS" 4
     fi
 else
-    compiler -b -d -f -g -j -l -L -t -R -U none -x -X ${TTCN_FILES} ${ASN1_FILES} 2>&1 3>&1 | tee build.log
+    compiler -d -e -f -g -l -L -t -R -U none -x -X ${TTCN_FILES} ${ASN1_FILES} 2>&1 3>&1 | tee build.log
     if [ "$?" == "1" ]
     then 
         f_exit "Failed to generate ATS source code" 6
@@ -133,18 +136,19 @@ fi
 # Sart ATS generation - Step 2
 # Create working variables
 CC_FILES=`find ../src -name '*.c*'`
+FWK_FILES=`find ${PATH_DEV_ITS}/framework/ -name '*.c*'`
 CFG_FILES=`find ../etc -name '*.cfg'`
 
 # Sart ATS generation - Step 3
 if [ "${OSTYPE}" == "cygwin" ]
 then
-    ttcn3_makefilegen.exe -d -f -g -m -R -U none -e Ats${ATS_NAME} ${TTCN_FILES} ${ASN1_FILES} ${CC_FILES} ${CFG_FILES} | tee --append build.log
+    ttcn3_makefilegen.exe -d -f -g -m -R -U none -e Ats${ATS_NAME} ${TTCN_FILES} ${ASN1_FILES} ${CC_FILES} ${FWK_FILES} ${CFG_FILES} | tee --append build.log
     if [ "$?" == "1" ]
     then
         f_exit "Failed to compile ATS" 5
     fi
 else
-    ttcn3_makefilegen -d -f -g -m -R -U none -e Ats${ATS_NAME} ${TTCN_FILES} ${ASN1_FILES} ${CC_FILES} ${CFG_FILES} | tee --append build.log
+    ttcn3_makefilegen -d -f -g -m -R -U none -e Ats${ATS_NAME} ${TTCN_FILES} ${ASN1_FILES} ${CC_FILES} ${FWK_FILES} ${CFG_FILES} | tee --append build.log
     if [ "$?" == "1" ]
     then
         f_exit "Failed to generate ATS source code" 7
@@ -154,28 +158,43 @@ fi
 # Remove port skeletons to use src/<port skeletons>
 for i in `ls ../include/*.hh`
 do
-    rm ./`basename $i`
+    if [ -f ./`basename $i` ]
+    then
+	      rm ./`basename $i`
+    fi
 done
 for i in `ls ../src/*.cc`
 do
-    rm ./`basename $i`
+    if [ -f ./`basename $i` ]
+    then
+	      rm ./`basename $i`
+    fi
 done
+
+# Check if Makefile was generated
+if [ ! -f ./Makefile ]
+then
+    f_exit "Failed to generate ATS source code" 8
+fi
 
 # Patch ATS generated files
 #./bin/patch.bash 2>&1 3>&1 | tee --append build.log
 # Add compiler/linker options
+# -DASN_DISABLE_OER_SUPPORT is required for CAMCodec and DENMCodec
 if [ "$1" == "prof" ]
 then
-    CXXFLAGS_DEBUG_MODE='s/-Wall/-pg -Wall -std=c++11/g'
-    LDFLAGS_DEBUG_MODE='s/LDFLAGS = /LDFLAGS = -pg /g'
+    CXXFLAGS_DEBUG_MODE='s/-Wall/-pg -Wall -std=c++11 -pthreads -fstack-check -fstack-protector -DASN_DISABLE_OER_SUPPORT/g'
+    LDFLAGS_DEBUG_MODE='s/LDFLAGS = /LDFLAGS = -pg -pthread -fstack-check -fstack-protector/g'
 else
-    CXXFLAGS_DEBUG_MODE='s/-Wall/-g -Wall -std=c++11/g'
-    LDFLAGS_DEBUG_MODE='s/LDFLAGS = /LDFLAGS = -g -L${PATH_DEV_ITS}/lib -lItsAsn /g'
+    CXXFLAGS_DEBUG_MODE='s/-Wall/-ggdb -O0 -Wall -std=c++11 -pthread -fstack-check -fstack-protector -DASN_DISABLE_OER_SUPPORT/g'
+    LDFLAGS_DEBUG_MODE='s/LDFLAGS = /LDFLAGS = -g -pthread -fstack-check -fstack-protector/g'
 fi
-ADD_INCLUDE='/CPPFLAGS = /a\\CPPFLAGS += -I../include -I../../LibIts/Common/include -I../../LibIts/BTP/include -I../../LibIts/DENM/include -I../../LibIts/DENM/include -I$(HOME_INC) -I.'
+ADD_INCLUDE='/CPPFLAGS = /a\\CPPFLAGS += -I/usr/local/share -I$(PATH_DEV_ITS)/include -I$(PATH_DEV_ITS)/include/asn1 -I$(PATH_DEV_ITS)/framework/include -I../include -I../../LibIts/Common/include -I../../LibIts/DENM/include -I$(HOME_INC) -I.'
+ADD_LIBRARIES='s/LINUX_LIBS = -lxml2/LINUX_LIBS = -lrt -lxml2 -lpcap -L$(PATH_DEV_ITS)\/lib -lItsAsn /g'
 sed --in-place "${CXXFLAGS_DEBUG_MODE}" ./Makefile 
 sed --in-place "${LDFLAGS_DEBUG_MODE}" ./Makefile
 sed --in-place "${ADD_INCLUDE}" ./Makefile
+sed --in-place "${ADD_LIBRARIES}" ./Makefile
 # Update clean clause
 CLEAN_LINE='s/$(RM) $(EXECUTABLE)/$(RM) ..\/bin\/$(EXECUTABLE) ..\/src\/*.o/g'
 sed --in-place "${CLEAN_LINE}" ./Makefile
@@ -189,11 +208,15 @@ ADD_PORT='/PLATFORM = /aPORT=12000'
 sed --in-place "${ADD_PORT}" ./Makefile
 sed --in-place "${ADD_HOST}" ./Makefile
 ADD_RUN_LINE_1='$arun: all'
-ADD_RUN_LINE_2='$a\\t@$(PWD)/../bin/$(EXECUTABLE) $(HOST) $(PORT)'
+ADD_RUN_LINE_2='$a\\t@sudo LD_LIBRARY_PATH=$(LD_LIBRARY_PATH) $(PWD)/../bin/$(EXECUTABLE) $(HOST) $(PORT)'
 sed --in-place "${ADD_RUN_LINE_1}" ./Makefile 
 sed --in-place "${ADD_RUN_LINE_2}" ./Makefile 
 ADD_RUN_LINE_1='$arun_d: all'
 ADD_RUN_LINE_2='$a\\t@gdb --args $(PWD)/../bin/$(EXECUTABLE) $(HOST) $(PORT)'
+sed --in-place "${ADD_RUN_LINE_1}" ./Makefile 
+sed --in-place "${ADD_RUN_LINE_2}" ./Makefile 
+ADD_RUN_LINE_1='$arun_v: all'
+ADD_RUN_LINE_2='$a\\t@sudo LD_LIBRARY_PATH=$(LD_LIBRARY_PATH) valgrind -v --tool=memcheck --leak-check=yes --show-reachable=yes --track-fds=yes --run-cxx-freeres=yes $(PWD)/../bin/$(EXECUTABLE) $(HOST) $(PORT)'
 sed --in-place "${ADD_RUN_LINE_1}" ./Makefile 
 sed --in-place "${ADD_RUN_LINE_2}" ./Makefile 
 
@@ -201,6 +224,7 @@ sed --in-place "${ADD_RUN_LINE_2}" ./Makefile
 make all 2>&1 3>&1 | tee --append build.log
 if [ "$?" == "1" ]
 then
-    f_exit "Failed to generate ATS source code" 8
+    f_exit "Failed to generate ATS source code" 9
 fi
+../bin/Ats${ATS_NAME} -v
 f_exit "Build done successfully" 0
